Improve LlamaGuard evaluation: reduce false unsafe classifications

+++ steps.py
@@ -33,7 +33,7 @@ from src.attacker import AdaptiveAttacker
 DATASET_PATH = "data/AdvBench_100.csv"
 N_GOALS = None
 RANDOM_SEED = 42
-PRECISION = "int4"
+PRECISION = "fp16"
 
 
 CHECKPOINT_PATH = f"results/checkpoint_{PRECISION}.json"
